reorg profiler files
