--- pytorch_lightning/utilities/__init__.py
@@ -1,1 +1,0 @@
-from .profiler import Profiler, AdvancedProfiler
